refactor and format document
