+++ routes/api/usersApi.js
@@ -3,17 +3,15 @@ const router = express.Router();
 const ctrlUser = require("../../controller/usersController");
 const authorization = require("../../tools/authorization");
 
-const path = require('path');
-const fs = require('fs').promises;
-const multer = require('multer');
-const uploadDir = path.join(process.cwd(), 'tmp');
-const storeImage = path.join(process.cwd(), 'images');
+const path = require("path");
+const multer = require("multer");
+const uploadDir = path.join(process.cwd(), "tmp");
 
 const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
+  destination: (_, __, cb) => {
     cb(null, uploadDir);
   },
-  filename: (req, file, cb) => {
+  filename: (_, file, cb) => {
     cb(null, file.originalname);
   },
   limits: {
@@ -29,6 +27,11 @@ router.post("/signup", ctrlUser.signUp);
 router.post("/login", ctrlUser.logIn);
 router.get("/logout", authorization, ctrlUser.logOut);
 router.get("/current", authorization, ctrlUser.current);
-router.patch("/avatars", authorization, upload.single('picture'), ctrlUser.avatar)
+router.patch(
+  "/avatars",
+  authorization,
+  upload.single("avatar"),
+  ctrlUser.avatar
+);
 
 module.exports = router;
